refactor(assets): Move files about transactions
