+++ lib/app/accounts/details/account_details.dart
@@ -7,10 +7,10 @@ import 'package:monekin/app/accounts/account_form.dart';
 import 'package:monekin/app/accounts/details/account_details_actions.dart';
 import 'package:monekin/app/assets/asset_details_page.dart';
 import 'package:monekin/app/layout/page_framework.dart';
-import 'package:monekin/app/transactions/label_value_info_list.dart';
-import 'package:monekin/app/transactions/transactions.page.dart';
-import 'package:monekin/app/transactions/widgets/transaction_list.dart';
-import 'package:monekin/app/transactions/widgets/transaction_list_tile.dart';
+import 'package:monekin/core/presentation/widgets/label_value_info_list.dart';
+import 'package:monekin/app/transactions/list/transactions.page.dart';
+import 'package:monekin/app/transactions/list/widgets/transaction_list.dart';
+import 'package:monekin/app/transactions/list/widgets/transaction_list_tile.dart';
 import 'package:monekin/core/database/services/account/account_service.dart';
 import 'package:monekin/core/database/services/account/investment_service.dart';
 import 'package:monekin/core/database/services/exchange-rate/exchange_rate_service.dart';
