+++ lib/app/debts/debt_details_page.dart
@@ -5,9 +5,9 @@ import 'package:intl/intl.dart';
 import 'package:monekin/app/debts/components/add_register_to_debt_modal.dart';
 import 'package:monekin/app/debts/debt_form_page.dart';
 import 'package:monekin/app/layout/page_framework.dart';
-import 'package:monekin/app/transactions/transaction_details.page.dart';
-import 'package:monekin/app/transactions/widgets/transaction_list.dart';
-import 'package:monekin/app/transactions/widgets/transaction_list_tile.dart';
+import 'package:monekin/app/transactions/details/transaction_details.page.dart';
+import 'package:monekin/app/transactions/list/widgets/transaction_list.dart';
+import 'package:monekin/app/transactions/list/widgets/transaction_list_tile.dart';
 import 'package:monekin/core/database/services/debts/debt_service.dart';
 import 'package:monekin/core/extensions/padding.extension.dart';
 import 'package:monekin/core/models/debt/debt.dart';
