+++ lib/app/stats/widgets/movements_distribution/category_stats_modal.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:monekin/app/stats/widgets/movements_distribution/tr_distribution_chart_item.dart';
-import 'package:monekin/app/transactions/transactions.page.dart';
+import 'package:monekin/app/transactions/list/transactions.page.dart';
 import 'package:monekin/core/extensions/color.extensions.dart';
 import 'package:monekin/core/models/category/category.dart';
 import 'package:monekin/core/models/supported-icon/icon_displayer.dart';
