+++ lib/app/transactions/details/utils/show_pay_modal.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:monekin/app/transactions/utils/transaction_details.utils.dart';
+import 'package:monekin/app/transactions/details/utils/transaction_details.utils.dart';
 import 'package:monekin/core/models/transaction/transaction.dart';
 import 'package:monekin/core/presentation/widgets/modal_container.dart';
 import 'package:monekin/core/routes/route_utils.dart';
