+++ lib/app/transactions/details/utils/transaction_details.utils.dart
