+++ lib/app/transactions/details/widgets/translucent_transaction_status_card.dart
